requirements.txt
azure-cognitiveservices-speech==1.34.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pyaudio==0.2.11
wave==0.0.2
colorama==0.4.6
logging
asyncio
